Added endpoints to admin controller and added utility function

+++ controllers/product_controllers.js
@@ -34,30 +34,26 @@ exports.getProduct = asyncHandler(async (req, res, next) => {
   })
 })
 
-// WILL NEED TO MOVE THIS ROUTE AND CHANGE ACCESS TO ADMIN. MOVE TO ADMIN CONTROLLER
 // @desc Create product
-// @route POST /api/v1/products <=== CHANGE TO ADMIN ONCE MOVED
-// access Admin
+// @route POST /api/v1/products
+// access  Private/Admin
 exports.createProducts = asyncHandler(async (req, res, next) => {
-  //
-  //   if (req.user.role !== 'admin') {
-  //     return next(new ErrorResponse(`User ${req.user.userId} not authorized`))
-  //   }
+  if (req.user.role !== 'admin' && process.env.NODE_ENV === 'production') {
+    return next(new ErrorResponse(`User ${req.user.userId} not authorized`))
+  }
 
   const product = await Products.create(req.body)
 
   res.status(201).json({ success: true, data: product })
 })
 
-// WILL NEED TO MOVE THIS ROUTE AND CHANGE ACCESS TO ADMIN. MOVE TO ADMIN CONTROLLER
 // @desc Update product
-// @route PUT /api/v1/products <=== CHANGE TO ADMIN ONCE MOVED
-// access Admin
+// @route PUT /api/v1/products
+// access Private/Admin
 exports.updateProduct = asyncHandler(async (req, res, next) => {
-  //
-  //   if (req.user.role !== 'admin') {
-  //     return next(new ErrorResponse(`User ${req.user.userId} not authorized`))
-  //   }
+  if (req.user.role !== 'admin' && process.env.NODE_ENV === 'production') {
+    return next(new ErrorResponse(`User ${req.user.userId} not authorized`))
+  }
 
   const product = await Products.update(req.body, {
     where: {
@@ -81,13 +77,12 @@ exports.updateProduct = asyncHandler(async (req, res, next) => {
 })
 
 // @desc Delete product
-// @route DELETE /api/v1/products <=== CHANGE TO ADMIN ONCE MOVED
-// access Admin
+// @route DELETE /api/v1/products
+// access  Private/Admin
 exports.deleteProduct = asyncHandler(async (req, res, next) => {
-  //
-  //   if (req.user.role !== 'admin') {
-  //     return next(new ErrorResponse(`User ${req.user.userId} not authorized`))
-  //   }
+  if (req.user.role !== 'admin' && process.env.NODE_ENV === 'production') {
+    return next(new ErrorResponse(`User ${req.user.userId} not authorized`))
+  }
   const product = await Products.destroy({
     where: {
       productId: req.params.productId,
